Move routines into shared.py
